# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bookstore E-commerce API
  version: 1.1.0
  description: REST API for authentication, catalog, cart, and order workflows.
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/books:
    get:
      tags: [Books]
      summary: List books
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, default: 10 }
        - in: query
          name: sortBy
          schema: { type: string, default: title }
        - in: query
          name: sortDir
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookPageResponse'
    post:
      tags: [Books]
      summary: Create book (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/orders/checkout:
    post:
      tags: [Orders]
      summary: Checkout cart
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        firstName: { type: string }
        lastName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: Backward-compatible alias of accessToken }
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, format: int64 }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string }
    BookDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        author: { type: string }
        isbn: { type: string }
        description: { type: string }
        price: { type: number }
        stockQuantity: { type: integer }
        imageUrl: { type: string }
        categoryId: { type: integer, format: int64 }
        categoryName: { type: string }
    BookPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookDTO'
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    CheckoutRequest:
      type: object
      required: [shippingAddress]
      properties:
        shippingAddress: { type: string }
    OrderDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        totalAmount: { type: number }
        status: { type: string }
        shippingAddress: { type: string }
        createdAt: { type: string, format: date-time }
    ApiErrorResponse:
      type: object
      properties:
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
        timestamp: { type: string, format: date-time }
        errors:
          type: object
          additionalProperties:
            type: string